Fixed the error ("Undefined function or method 'prod' for input arguments of type 'int32'.") when calling prod with an int32 argument for older versions of Matlab (thanks to Nikolai Lebedev).
Fixed the error ("Undefined function or method 'prod' for input
arguments of type 'int32'.") when calling prod with an int32 argument
for older versions of Matlab (thanks to Nikolai Lebedev).

+++ UGM/sub/UGM_makeEdgeStruct.m
@@ -34,7 +34,7 @@ edgeStruct.nEdges = size(edgeEnds,1);
 
 % Handle other arguments
 if isscalar(nStates)
-   nStates = repmat(nStates,[nNodes 1]);
+   nStates = repmat(nStates,[double(nNodes) 1]);
 end
 edgeStruct.nStates = int32(nStates(:));
 edgeStruct.useMex = useMex;
